+++ tests/TestCase/Utility/BackupExportTest.php
@@ -41,7 +41,7 @@ class BackupExportTest extends TestCase
      * Called before every test method
      * @return void
      */
-    protected function setUp(): void
+    public function setUp(): void
     {
         parent::setUp();