+++ tests/TestCase/Utility/BackupImportTest.php
@@ -43,7 +43,7 @@ class BackupImportTest extends TestCase
      * Called before every test method
      * @return void
      */
-    protected function setUp(): void
+    public function setUp(): void
     {
         parent::setUp();